update jamonapi.com link
I would hope you agree this is an appropriate change to make, but of course if you prefer to link to some other resource, feel free. 

I see that the old domain must have been given up some years ago, as I see the current site content (unrelated to this project) has been like that for a few years, per web.archive.org, and I see that in 2015 it was redirecting instead to the sourceforge page.

I realize it's possible that page is no longer being kept up to date, but it seemed better to propose that here rather than leave the old link, going where it did. And FWIW, that soureforge page has a itself two references to jamonapi.com, one being an email address (which I would assume doesn't work any more, but again Steve you will know best).

And thanks for all your years of effort with the tool (and your updating things here even in 2021!)